Add files via upload
Overloaded weapon to allow for constructor to initialize spreading behavior (i.e. shotgun pattern). Also added string name member variable for weapons. No constructor is implemented with string name, but this is more easily fixed when said member variable is necessary for ui.

+++ bullet.cpp
@@ -2,7 +2,7 @@
 #include <QGraphicsScene>
 #include <QDebug>
 
-Bullet::Bullet(bool playerMade, int x, int y)
+Bullet::Bullet(bool playerMade, double x, double y)
     :Actor(-2,-2,2)
 {
     movementTimer = new QTimer();
