+++ bullet.h
@@ -11,10 +11,10 @@ class Bullet : public QObject, public Actor
 private:
     QTimer* movementTimer;
     bool playerMade;
-    int xMove;
-    int yMove;
+    double xMove;
+    double yMove;
 public:
-    Bullet(bool playerMade, int xMove, int yMove);
+    Bullet(bool playerMade, double xMove, double yMove);
 
 public slots:
     void move();
